+++ metatensor-torch/include/metatensor/torch/atomistic/model.hpp
@@ -303,15 +303,12 @@ METATENSOR_TORCH_EXPORT void load_model_extensions(
 
 /// Check and then load the metatensor atomistic model at the given `path`.
 ///
-/// The model is returned on the given `device`.
-///
 /// This function calls `check_atomistic_model(path)` and
 /// `load_model_extensions(path, extension_directory)` before attempting to load
 /// the model.
 METATENSOR_TORCH_EXPORT torch::jit::Module load_atomistic_model(
     std::string path,
-    c10::optional<std::string> extensions_directory,
-    c10::optional<c10::Device> device = c10::nullopt
+    c10::optional<std::string> extensions_directory = c10::nullopt
 );
 
 /// Get the multiplicative conversion factor to use to convert from unit `from`