+++ metatensor-torch/CHANGELOG.md
@@ -23,6 +23,8 @@ a changelog](https://keepachangelog.com/en/1.1.0/) format. This project follows
 
 - `ModelCapabilities::dtype`, used by the model to communicate the dtype it
   wants to use for inputs and outputs.
+- The `load_model_extensions()` function to facilitate loading a model using
+  TorchScript extensions
 
 ### metatensor-torch Python
 
@@ -30,10 +32,10 @@ a changelog](https://keepachangelog.com/en/1.1.0/) format. This project follows
 
 - `ModelCapabilities.dtype`, used by the model to communicate the dtype it
   wants to use for inputs and outputs.
-
 - The `device` that should be used by a model inside the ASE's
   `MetatensorCalculator` can now be specified by the user.
-
+- The `load_model_extensions()` and `load_atomistic_model` functions to
+  facilitate loading a model using TorchScript extensions
 
 ## [Version 0.3.0](https://github.com/lab-cosmo/metatensor/releases/tag/metatensor-torch-v0.3.0) - 2024-03-01